Change zustand to slice pattern for async storage
