+++ packages/app/features/home/Clock.tsx
@@ -1,4 +1,4 @@
-import { useClockStore, useInterval } from 'app/zustand'
+import { useBoundStore, useInterval } from 'app/zustand'
 import { H2 } from '@my/ui'
 
 const formatTime = (time: number) => {
@@ -7,8 +7,8 @@ const formatTime = (time: number) => {
 }
 
 export function Clock() {
-  const lastUpdate = useClockStore((state) => state.lastUpdate)
-  const tick = useClockStore((state) => state.tick)
+  const lastUpdate = useBoundStore((state) => state.lastUpdate)
+  const tick = useBoundStore((state) => state.tick)
 
   // Tick the time every second
   useInterval(() => {
