+++ packages/app/provider/Provider.tsx
@@ -1,10 +1,10 @@
 import { ToastViewport } from './ToastViewport'
 import config from 'app/tamagui.config'
-import { useThemeStore } from 'app/zustand'
+import { useBoundStore } from 'app/zustand'
 import { CustomToast, TamaguiProvider, TamaguiProviderProps, ToastProvider } from '@my/ui'
 
 export function Provider({ children, ...rest }: Omit<TamaguiProviderProps, 'config'>) {
-  const theme = useThemeStore((state) => state.theme)
+  const theme = useBoundStore((state) => state.theme)
 
   return (
     <TamaguiProvider config={config} defaultTheme={theme} disableInjectCSS {...rest}>
